backend/requirements.txt
pycountry==20.7.3
django==3.1
djangorestframework==3.11.1
django-cors-headers==3.4.0
pytz==2020.1
faker==4.1.2
psycopg2-binary==2.8.5
gunicorn==20.0.4
gevent==20.9.0
django-environ==0.4.5
social-auth-app-django==4.0.0